Sources/Unified/requirements.txt
jsonschema==4.21.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0